requirements.txt
autopep8==1.5.7
Django==2.2.19
django-debug-toolbar==3.2.1
django-multiselectfield==0.1.12
django-taggit==1.3.0
djangorestframework==3.12.4
isort==5.8.0
pdfkit==0.6.0
Pillow==8.2.0
pycodestyle==2.7.0
pytz==2021.1
reportlab==3.5.67
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.2
